# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "bluebell项目接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8084",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/admin/articles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员获取所有文章列表，支持状态筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "管理员获取文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "all",
                            "published",
                            "draft",
                            "offline"
                        ],
                        "type": "string",
                        "description": "文章状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "作者用户名",
                        "name": "author",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseArticleList"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/articles/{id}/featured": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员设置文章的精选状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员"
                ],
                "summary": "管理员设置文章精选",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "精选状态",
                        "name": "featured",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamArticleFeatured"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/article-stats/daily": {
            "get": {
                "description": "获取指定文章的日访问量统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "获取文章日访问量统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "article_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 90,
                        "minimum": 1,
                        "type": "integer",
                        "default": 7,
                        "description": "统计天数",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "日访问统计数据",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseArticleDailyStats"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/articles": {
            "get": {
                "description": "获取公开的文章列表，支持分页和状态筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "published",
                            "draft",
                            "offline"
                        ],
                        "type": "string",
                        "default": "published",
                        "description": "文章状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "time",
                            "hot"
                        ],
                        "type": "string",
                        "default": "time",
                        "description": "排序方式",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "栏目ID",
                        "name": "category_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseArticleList"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "作者创建新文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "创建文章",
                "parameters": [
                    {
                        "description": "文章参数",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamCreateArticle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseCreateArticle"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/articles/featured": {
            "get": {
                "description": "获取精选文章列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章"
                ],
                "summary": "获取精选文章",
                "parameters": [
                    {
                        "maximum": 10,
                        "minimum": 1,
                        "type": "integer",
                        "default": 3,
                        "description": "数量限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "精选文章列表",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseFeaturedArticles"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/articles/view": {
            "post": {
                "description": "记录文章访问，带防刷机制（同一IP每天最多10次）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "记录文章访问（防刷）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "article_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "记录成功",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/controller._ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller._ResponseArticleDailyStats": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
...
        "controller._ResponseArticleList": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "list": {},
                        "page": {
                            "type": "integer"
                        },
                        "pages": {
                            "type": "integer"
                        },
                        "size": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {}
            }
        },
...
        "controller._ResponseCreateArticle": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
        "controller._ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "msg": {}
            }
        },
...
        "controller._ResponseFeaturedArticles": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
...
        "controller._ResponseSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/controller.ResCode"
                },
                "data": {},
                "msg": {}
            }
        },
...
        "models.ParamArticleFeatured": {
            "type": "object",
            "properties": {
                "is_featured": {
                    "type": "boolean"
                }
            }
        },
...
    }
}